refactor: split render into modules
